+++ src/game/test/TestPaintable.java
@@ -25,7 +25,7 @@ public class TestPaintable extends BasicGameObject implements Paintable {
 
 	@Override
 	public void paint(Graphics2D g2d) {
-		g2d.fillRect((int)x, (int)y, width, height);
+		g2d.fillRect((int)x, (int)y, (int)width, (int)height);
 	}
 
 	@Override
